Refactor package object to a normal object
