+++ caps-core/src/test/scala/org/opencypher/caps/api/record/RecordHeaderTest.scala
@@ -15,13 +15,13 @@
  */
 package org.opencypher.caps.api.record
 
-import org.opencypher.caps.ir.api.expr._
 import org.opencypher.caps.api.schema.Schema
 import org.opencypher.caps.api.schema.Schema.NoLabel
 import org.opencypher.caps.api.types._
 import org.opencypher.caps.impl.record._
 import org.opencypher.caps.impl.syntax.RecordHeaderSyntax._
-import org.opencypher.caps.impl.util.toVar
+import org.opencypher.caps.impl.util.VarConverters.toVar
+import org.opencypher.caps.ir.api.expr._
 import org.opencypher.caps.ir.api.{Label, PropertyKey, RelType}
 import org.opencypher.caps.test.BaseTestSuite
